+++ facebook_test.go
@@ -12,9 +12,9 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
+	"strings"
 	"testing"
 	"time"
-	"strings"
 )
 
 const (